backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-cloud-bigquery==3.14.1
python-dotenv==1.0.0
pydantic==2.5.3
pandas==2.1.4
